Mgc: nested graphs finished <version>jsk.17.2.0</version> next: <version>jsk.17.2.1</version>

+++ jx-modules/jx-utils/src/main/java/sk/utils/async/AtomicNotifier.java
@@ -28,7 +28,7 @@ import java.util.concurrent.atomic.AtomicLong;
 @AllArgsConstructor
 public class AtomicNotifier {
     final private AtomicLong al = new AtomicLong();
-    final private int maxValue;
+    final private long maxValue;
     final private long divider;
     final private C1<String> notifier;
